update move kt to view file
