--- app/src/main/java/ie/ul/frankscafe/Home.kt
@@ -1,4 +1,0 @@
-package ie.ul.frankscafe
-
-class Home {
-}
