--- app/src/main/java/ie/ul/frankscafe/Ordering.kt
@@ -1,4 +1,0 @@
-package ie.ul.frankscafe
-
-class Ordering {
-}
