+++ app/src/main/java/ie/ul/frankscafe/View/Home.kt
@@ -1,0 +1,4 @@
+package ie.ul.frankscafe.View
+
+class Home {
+}
