+++ app/src/main/java/ie/ul/frankscafe/View/Signin.kt
@@ -1,6 +1,7 @@
-package ie.ul.frankscafe
+package ie.ul.frankscafe.View
 
 import androidx.fragment.app.Fragment
+import ie.ul.frankscafe.R
 
 class Signin : Fragment(R.layout.signin) {
 }
